refactor(dashboard): move out of src/interface-adapters into top-level src/dashboard
The Phase 1 reorganization moved every BC concern into src/modules/<bc>/,
leaving src/interface-adapters/ with only the dashboard JS humble objects
(24 files). The dashboard is cross-BC — it renders data from every
bounded context — so placing it under a single module would lie about
ownership. Top-level src/dashboard/ matches the screaming-architecture
convention used for the other non-BC concerns (src/main/, src/frameworks/,
src/mcp/, etc.).

Moves:
- src/interface-adapters/views/dashboard/   → src/dashboard/
- src/tests/units/interface-adapters/views/dashboard/modules/
  → src/tests/units/dashboard/modules/

Updated references (paths and @/ aliases):
- scripts/copyAssets.mjs (build asset copy)
- src/main/routes.ts (@fastify/static root)
- 18 dashboard module test files (import paths)
- src/tests/acceptance/dashboardModulesCoverage.acceptance.test.ts
  (scanned directories)

Side benefit: drops src/interface-adapters/ entirely (now empty).
Top-level src/ no longer carries the legacy by-layer interface-adapters/
directory at all — every BC owns its own interface-adapters/ subtree
under src/modules/<bc>/.

Also fixes the longstanding preexisting 'join' unused warning in
dashboardModulesCoverage.acceptance.test.ts — the unused import was
the entry that flagged the typecheck since master; cleaning it as
part of the path rewrite reduces the typecheck noise to zero.

yarn typecheck: ZERO errors (first time since the Phase 1 reorg started)
yarn test:ci: 1379/1381 passing, 2 skipped, 0 failing
