+++ src/dashboard/modules/collapsibleList.js
