+++ src/dashboard/modules/icons.js
