+++ src/dashboard/modules/insightsReport.js
