+++ src/dashboard/modules/sessionMetrics.js
