+++ src/dashboard/modules/sharedViewHelpers.js
