+++ src/dashboard/modules/statsCharts.js
