+++ src/tests/acceptance/dashboardModulesCoverage.acceptance.test.ts
@@ -1,15 +1,15 @@
 import { describe, it, expect } from 'vitest';
 import { readdirSync, existsSync } from 'node:fs';
-import { join, resolve } from 'node:path';
+import { resolve } from 'node:path';
 
 const MODULES_DIR = resolve(
   process.cwd(),
-  'src/interface-adapters/views/dashboard/modules'
+  'src/dashboard/modules'
 );
 
 const TESTS_DIR = resolve(
   process.cwd(),
-  'src/tests/units/interface-adapters/views/dashboard/modules'
+  'src/tests/units/dashboard/modules'
 );
 
 function listJsModules(): string[] {
