+++ src/tests/units/dashboard/modules/formatting.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach } from 'vitest';
-import { formatTime, formatDuration, formatPhase, formatLogTime } from '@/interface-adapters/views/dashboard/modules/formatting.js';
-import { setLanguage } from '@/interface-adapters/views/dashboard/modules/i18n.js';
+import { formatTime, formatDuration, formatPhase, formatLogTime } from '@/dashboard/modules/formatting.js';
+import { setLanguage } from '@/dashboard/modules/i18n.js';
 
 describe('formatTime', () => {
   beforeEach(() => {
